Rename useTestRunSuspends.ts -> useTestRunsSuspends.ts
