+++ src/ui/components/Library/Team/View/TestRuns/TestRunListPanel/TestRunList.tsx
@@ -5,7 +5,7 @@ import { FixedSizeList, ListChildComponentProps } from "react-window";
 import { TestRun } from "shared/test-suites/TestRun";
 
 import { TestSuitePanelMessage } from "../../TestSuitePanelMessage";
-import { useTestRunsSuspends } from "../hooks/useTestRunSuspends";
+import { useTestRunsSuspends } from "../hooks/useTestRunsSuspends";
 import { TestRunsContext } from "../TestRunsContextRoot";
 import { TestRunListItem } from "./TestRunListItem";
 import styles from "./TestRunList.module.css";
